Add files via upload
In the process of understanding and studying Deeplearning, I am using a text book called Deep Learning with Python. And this is the most basic example to start with. It works with and IMDB dataset to predict sentiments into positive and negative. Making it a good study on binary classification.